@@ -1,13 +1,13 @@
 name: api_repository
 description: Dart package that manages HTTP API calls for JogabiliApp
-version: 1.0.0
+version: 1.0.1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   meta: ^1.2.4
-  equatable: ^2.0.2
-  http: ^0.13.1
+  equatable: ^2.0.3
+  http: ^0.13.4
