@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -29,7 +29,8 @@ dependencies:
   marquee: ^2.1.0
   palette_generator: ^0.3.0
   url_launcher: ^6.0.9
-  just_audio: ^0.9.5
+  just_audio: ^0.9.18
+  audio_service: ^0.18.2
   api_repository: 
     path: ./packages/api_repository
   episodes_repository:
