@@ -9,9 +9,7 @@ part 'episodes_event.dart';
 part 'episodes_state.dart';
 
 class EpisodesBloc extends Bloc<EpisodesEvent, EpisodesState> {
-  EpisodesBloc({@required this.episodesRepository})
-      : assert(episodesRepository != null),
-        super(EpisodesInitial());
+  EpisodesBloc({required this.episodesRepository}) : super(EpisodesInitial());
 
   final EpisodesRepository episodesRepository;
 
