@@ -4,11 +4,14 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:jogabili_app/blocs/player/player_bloc.dart';
 import 'package:jogabili_app/ui/constants/text_styles.dart';
 import 'package:marquee/marquee.dart';
-import 'package:palette_generator/palette_generator.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class PlayerPage extends StatefulWidget {
-  const PlayerPage({Key key, this.episode, this.playerBloc, this.bgColor})
+  const PlayerPage(
+      {Key? key,
+      required this.episode,
+      required this.playerBloc,
+      required this.bgColor})
       : super(key: key);
   final Episode episode;
   final PlayerBloc playerBloc;
@@ -26,8 +29,8 @@ class _PlayerPageState extends State<PlayerPage> {
   static final Color darkColor = Color(0xFFFFFFFF);
   final Color bgColor;
   Color colorForTheme = lightColor;
-  Color reverseColorForTheme;
-  Brightness appBarBrightness;
+  Color? reverseColorForTheme;
+  Brightness? appBarBrightness;
   double sliderValue = 5;
 
   @override
@@ -37,11 +40,6 @@ class _PlayerPageState extends State<PlayerPage> {
   }
 
   void setBgColor() async {
-    Color newBgColor = (await PaletteGenerator.fromImageProvider(
-            NetworkImage(episode.imageUrl)))
-        .dominantColor
-        .color;
-
     setState(() {
       colorForTheme =
           ThemeData.estimateBrightnessForColor(bgColor) == Brightness.light
@@ -72,7 +70,7 @@ class _PlayerPageState extends State<PlayerPage> {
           title: Container(
             height: 50,
             child: Marquee(
-              text: episode.title,
+              text: episode.title!,
               style: TextStyles.episodeTitlePlayerTextStyle(colorForTheme),
               blankSpace: 20,
               startAfter: Duration(seconds: 3),
@@ -95,7 +93,7 @@ class _PlayerPageState extends State<PlayerPage> {
           child: Column(
             children: [
               Image(
-                image: NetworkImage(episode.imageUrl),
+                image: NetworkImage(episode.imageUrl!),
               ),
               Padding(
                 padding: const EdgeInsets.only(top: 50, bottom: 15),
@@ -111,7 +109,7 @@ class _PlayerPageState extends State<PlayerPage> {
                           builder: (context, snapshot) {
                             if (snapshot.hasData)
                               return Slider(
-                                value: snapshot.data.inSeconds.toDouble(),
+                                value: snapshot.data!.inSeconds.toDouble(),
                                 min: 0,
                                 max: state.duration.inSeconds.toDouble(),
                                 divisions:
@@ -181,7 +179,7 @@ class _PlayerPageState extends State<PlayerPage> {
       Container(
           child: Center(
         child: SingleChildScrollView(
-          child: Text(episode.longDescription,
+          child: Text(episode.longDescription!,
               style: TextStyles.episodeLongDescriptionTextStyle(colorForTheme)),
         ),
       )),
@@ -196,14 +194,14 @@ class _PlayerPageState extends State<PlayerPage> {
       ));
       content.add(Container(
           child: ListView.builder(
-        itemCount: episode.links.length,
+        itemCount: episode.links!.length,
         itemBuilder: (context, index) {
           return TextButton(
               onPressed: () {
-                launch(episode.links[index].linkUrl);
+                launch(episode.links![index].linkUrl!);
               },
               child: Text(
-                episode.links[index].title,
+                episode.links![index].title!,
                 textAlign: TextAlign.start,
               ));
         },
@@ -224,13 +222,14 @@ class _PlayerPageState extends State<PlayerPage> {
               return ListTile(
                   onTap: () {
                     context.read<PlayerBloc>().add(PlayerJump(
-                        timestampToDuration(episode.blocks[index].timeStamp)));
+                        timestampToDuration(
+                            episode.blocks![index].timeStamp!)));
                   },
                   title: Container(
                     height: 50,
                     child: Marquee(
                       text:
-                          "${episode.blocks[index].title} - ${episode.blocks[index].timeStamp} ",
+                          "${episode.blocks![index].title} - ${episode.blocks![index].timeStamp} ",
                       style: TextStyles.episodeTabTextStyle(colorForTheme),
                       blankSpace: 20,
                       startAfter: Duration(seconds: 3),
@@ -242,7 +241,7 @@ class _PlayerPageState extends State<PlayerPage> {
             separatorBuilder: (context, index) => Divider(
                   color: colorForTheme,
                 ),
-            itemCount: episode.blocks.length),
+            itemCount: episode.blocks!.length),
       )));
     }
 
@@ -275,11 +274,11 @@ class _PlayerPageState extends State<PlayerPage> {
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: [
                   Text(
-                    "${snapshot.data.inHours}:${fullNumber(snapshot.data.inMinutes.remainder(60))}:${fullNumber(snapshot.data.inSeconds.remainder(60))}",
+                    "${snapshot.data!.inHours}:${fullNumber(snapshot.data!.inMinutes.remainder(60))}:${fullNumber(snapshot.data!.inSeconds.remainder(60))}",
                     style: TextStyles.episodeLengthTextStyle(colorForTheme),
                   ),
                   Text(
-                    "${duration.inHours}:${duration.inMinutes.remainder(60)}:${(duration.inSeconds.remainder(60))}",
+                    "${duration.inHours}:${fullNumber(duration.inMinutes.remainder(60))}:${fullNumber(duration.inSeconds.remainder(60))}",
                     style: TextStyles.episodeLengthTextStyle(colorForTheme),
                   )
                 ],
@@ -338,7 +337,7 @@ class _PlayerPageState extends State<PlayerPage> {
                       child: IconButton(
                           onPressed: () {
                             context.read<PlayerBloc>().add(PlayerJump(Duration(
-                                seconds: snapshot.data.inSeconds - 5)));
+                                seconds: snapshot.data!.inSeconds - 5)));
                           },
                           icon: Icon(
                             Icons.fast_rewind,
@@ -355,14 +354,21 @@ class _PlayerPageState extends State<PlayerPage> {
                             color: colorForTheme,
                             borderRadius:
                                 BorderRadius.all(Radius.circular(50))),
-                        child: IconButton(
-                            onPressed: () {
-                              context.read<PlayerBloc>().add(PlayerSwitch());
-                            },
-                            icon: Icon(
-                              state.isPlaying ? Icons.pause : Icons.play_arrow,
-                              color: reverseColorForTheme,
-                            )),
+                        child: StreamBuilder<bool>(
+                          stream: state.playingStream,
+                          builder: (context, snapshot) => IconButton(
+                              onPressed: () {
+                                context.read<PlayerBloc>().add(PlayerSwitch());
+                              },
+                              icon: Icon(
+                                snapshot.data == null
+                                    ? Icons.play_arrow
+                                    : snapshot.data!
+                                        ? Icons.pause
+                                        : Icons.play_arrow,
+                                color: reverseColorForTheme,
+                              )),
+                        ),
                       ),
                     ),
                     Padding(
@@ -370,7 +376,7 @@ class _PlayerPageState extends State<PlayerPage> {
                       child: IconButton(
                           onPressed: () {
                             context.read<PlayerBloc>().add(PlayerJump(Duration(
-                                seconds: snapshot.data.inSeconds + 5)));
+                                seconds: snapshot.data!.inSeconds + 5)));
                           },
                           icon: Icon(
                             Icons.fast_forward,
