@@ -17,26 +17,28 @@ class PlayerPlaying extends PlayerState {
   final Stream<Duration> position;
   final Duration duration;
   final bool isPlaying;
+  final Stream<bool> playingStream;
   final Color bgColor;
   final Episode episode;
-  PlayerPlaying(
-      this.position, this.duration, this.isPlaying, this.bgColor, this.episode);
+  PlayerPlaying(this.position, this.duration, this.playingStream,
+      this.isPlaying, this.bgColor, this.episode);
 
   PlayerPlaying copyWith(
-      {Stream<Duration> position,
-      Duration duration,
-      bool isPlaying,
-      Color color,
-      Episode episode}) {
+      {Stream<Duration>? position,
+      Duration? duration,
+      Stream<bool>? playingStream,
+      bool? isPlaying,
+      Color? color,
+      Episode? episode}) {
     return PlayerPlaying(
         position ?? this.position,
         duration ?? this.duration,
+        playingStream ?? this.playingStream,
         isPlaying ?? this.isPlaying,
         color ?? this.bgColor,
         episode ?? this.episode);
   }
 
   @override
-  // TODO: implement props
   List<Object> get props => [position, duration, isPlaying, bgColor, episode];
 }
